backend/requirements.txt
Django>=5.2
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers
